+++ README.md
@@ -7,5 +7,40 @@ I used Android Material design standards and made sure it uses as many transitio
 
 Make sure to tap on the market orders, because they open to give you more info, like range and location.
 
-# Future
-I have also been working on adding it a "bot" feature that tracks specific items in specific regions of space and notify you when the price of buy or sell orders pass your threshold. Along with that I added network caching so you can still use quite a bit of it in offline mode, and working on building it into the database so it if quicker and requires internet only for specific calls to update info or ones not stored.
+
+
+# Intent Integration
+
+Courtesy of EvaNova, here is the class needed to open MarketBot from any app that you would like:
+
+    import android.content.Context;
+    import android.content.Intent;
+    import android.content.pm.PackageManager;
+    import android.content.pm.ResolveInfo;
+    ​
+    import java.util.List;
+    ​
+    public class MarketBot {
+        private MarketBot() {}
+    ​
+        public static void launch(
+                final Context context,
+                final long itemID,
+                final long regionID) {
+            final Intent intent = new Intent("com.w9jds.marketbot.intent.action.OPEN_TYPE");
+            intent.putExtra("typeId", itemID);
+            intent.putExtra("regionId", regionID);
+            context.startActivity(intent);
+        }
+    ​
+        public static boolean isAvailable(final Context context) {
+            final Intent intent = new Intent("com.w9jds.marketbot.intent.action.OPEN_TYPE");
+            return isIntentAvailable(context, intent);
+        }
+    ​
+        private static boolean isIntentAvailable(Context ctx, Intent intent) {
+            final PackageManager mgr = ctx.getPackageManager();
+            List<ResolveInfo> list = mgr.queryIntentActivities(intent, 0);
+            return list.size() > 0;
+        }
+    }
