+++ app/build.gradle
@@ -8,8 +8,8 @@ android {
     defaultConfig {
         applicationId "com.w9jds.marketbot"
         minSdkVersion 21
-        versionCode 6
-        versionName "v1.0.1"
+        versionCode 8
+        versionName "v1.0.3"
     }
     buildTypes {
         release {
