+++ app/src/main/java/com/w9jds/marketbot/classes/MarketBot.java
@@ -32,4 +32,5 @@ public class MarketBot extends Application {
                 .storageModule(new StorageModule())
                 .build();
     }
+
 }
