+++ app/src/main/java/com/w9jds/marketbot/classes/components/StorageComponent.java
@@ -1,9 +1,12 @@
 package com.w9jds.marketbot.classes.components;
 
 import com.w9jds.marketbot.activities.InfoActivity;
+import com.w9jds.marketbot.activities.ItemActivity;
+import com.w9jds.marketbot.activities.MainActivity;
 import com.w9jds.marketbot.classes.StorageScope;
 import com.w9jds.marketbot.classes.modules.StorageModule;
-import com.w9jds.marketbot.data.DataManager;
+import com.w9jds.marketbot.data.BaseDataManager;
+import com.w9jds.marketbot.data.loader.DataManager;
 
 import dagger.Component;
 
@@ -16,5 +19,6 @@ public interface StorageComponent {
 
     void inject(DataManager manager);
     void inject(InfoActivity infoActivity);
+    void inject(ItemActivity itemActivity);
 
 }
