+++ app/src/main/java/com/w9jds/marketbot/classes/modules/ApplicationModule.java
@@ -2,8 +2,13 @@ package com.w9jds.marketbot.classes.modules;
 
 import android.app.Application;
 import android.content.SharedPreferences;
+import android.database.sqlite.SQLiteDatabase;
 import android.preference.PreferenceManager;
 
+import com.w9jds.marketbot.classes.StorageScope;
+import com.w9jds.marketbot.data.Database;
+
+import javax.inject.Named;
 import javax.inject.Singleton;
 
 import dagger.Module;
@@ -12,16 +17,16 @@ import dagger.Provides;
 @Module
 public class ApplicationModule {
 
-    private Application appModule;
+    private Application application;
 
     public ApplicationModule(Application application) {
-        appModule = application;
+        this.application = application;
     }
 
     @Provides
     @Singleton
-    Application privideApplication() {
-        return appModule;
+    Application provideApplication() {
+        return application;
     }
 
     @Provides
@@ -29,17 +34,5 @@ public class ApplicationModule {
     SharedPreferences provideSharedPreferences(Application application) {
         return PreferenceManager.getDefaultSharedPreferences(application);
     }
-
-//    @Provides @Named("write")
-//    @StorageScope
-//    public SQLiteDatabase provideWritableDatabase() {
-//        return new Database(context).getWritableDatabaseHelper();
-//    }
-//
-//    @Provides @Named("read")
-//    @StorageScope
-//    public SQLiteDatabase provideReadableDatabase() {
-//        return new Database(context).getReadableDatabaseHelper();
-//    }
 
 }
