+++ app/src/main/java/com/w9jds/marketbot/data/DataLoadingSubject.java
@@ -15,6 +15,7 @@ public interface DataLoadingSubject {
     interface DataLoadingCallbacks {
         void dataStartedLoading();
         void dataFinishedLoading();
+        void dataFailedLoading(String errorMessage);
     }
 
     interface DataUpdatingCallbacks {
