+++ app/src/main/java/com/w9jds/marketbot/data/models/Bot.java
@@ -3,7 +3,7 @@ package com.w9jds.marketbot.data.models;
 import android.os.PersistableBundle;
 
 import com.w9jds.marketbot.classes.StorageColumn;
-import com.w9jds.marketbot.data.storage.DataContracts.BotEntry;
+import com.w9jds.marketbot.data.storage.BotEntry;
 
 /**
  * Created by Jeremy on 3/5/2016.
