+++ app/src/main/java/com/w9jds/marketbot/ui/fragments/TypeInfoTab.java
@@ -16,7 +16,7 @@ import com.w9jds.eveapi.Models.TypeInfo;
 import com.w9jds.marketbot.R;
 import com.w9jds.marketbot.activities.ItemActivity;
 import com.w9jds.marketbot.data.BaseDataManager;
-import com.w9jds.marketbot.data.DataManager;
+import com.w9jds.marketbot.data.loader.DataManager;
 
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
@@ -89,6 +89,11 @@ public final class TypeInfoTab extends Fragment implements BaseDataManager.DataL
 
         dataManager = new DataManager(getActivity().getApplication()) {
             @Override
+            public void onProgressUpdate(int page, int totalPages) {
+
+            }
+
+            @Override
             public void onDataLoaded(List<? extends MarketItemBase> data) {
 
             }
@@ -148,6 +153,11 @@ public final class TypeInfoTab extends Fragment implements BaseDataManager.DataL
 
         mass.setVisibility(View.VISIBLE);
         capacity.setVisibility(View.VISIBLE);
+    }
+
+    @Override
+    public void dataFailedLoading(String errorMessage) {
+
     }
 
 }
